[Clean Code] Move steps definitions to the store.
