+++ src/components/pages/steps.tsx
@@ -6,67 +6,17 @@ import {useHistory} from 'react-router'
 
 import {FirebaseAuth} from 'database/firebase'
 import {RootState} from 'store/actions'
-import {joinList, prepareT} from 'store/i18n'
+import {joinList} from 'store/i18n'
 import {useProject} from 'store/selections'
-import {StepId} from 'store/steps'
+import Steps from 'store/steps'
 import {Page, getPath} from 'store/url'
 
 import Button from 'components/button'
 import Layout from 'components/layout'
 import CreateAccountPopup from 'components/create_account_popup'
 import Step from 'components/step'
-import competencesIcon from 'images/competences-ico.svg'
-import definitionIcon from 'images/definition-ico.svg'
-import formationIcon from 'images/formation-ico.svg'
 
 const soonAvailable = (): void => window.alert('Bientôt disponible...')
-
-export interface StepInfo {
-  color: string
-  icon: string
-  isOpen?: boolean
-  isLastStep?: boolean
-  page: Page
-  shortTitle?: string
-  stepId: StepId
-  title: string
-}
-
-// TODO(cyrille): Move to store.
-export const STEPS: readonly StepInfo[] = [
-  {
-    color: colors.LIGHT_TAN,
-    icon: definitionIcon,
-    page: ['DEFINITION'],
-    shortTitle: prepareT('Définition'),
-    stepId: 'definition',
-    title: prepareT('Définition de votre projet'),
-  },
-  {
-    color: colors.SILVER,
-    icon: competencesIcon,
-    page: ['SKILLS'],
-    stepId: 'skills',
-    title: prepareT('Compétences'),
-  },
-  {
-    color: colors.LIGHT_SKY_BLUE,
-    icon: formationIcon,
-    page: ['TRAINING'],
-    shortTitle: prepareT('Formation'),
-    stepId: 'training',
-    title: prepareT('Formations'),
-  },
-  {
-    color: colors.LIGHT_SKY_BLUE,
-    icon: formationIcon, // TODO(émilie): Update when known
-    isLastStep: true,
-    page: ['TRAINING'], // TODO(émilie): Updated when done
-    shortTitle: prepareT('Entretiens'),
-    stepId: 'interview',
-    title: prepareT('Préparer un entretien'),
-  },
-]
 
 const stepsStyle: React.CSSProperties = {
   padding: '30px 0',
@@ -101,7 +51,7 @@ const StepsPage = (): React.ReactElement => {
   const isConnected = (currentUser !== null && name !== undefined && lastName !== undefined)
   const project = useProject()
 
-  const nextStep = STEPS.find(({stepId}) => !project.steps?.[stepId]?.completed)?.stepId
+  const nextStep = Steps.find(({stepId}) => !project.steps?.[stepId]?.completed)?.stepId
   const nextStepRef = useRef<HTMLDivElement>(null)
   useEffect((): (() => void) => {
     // TODO(cyrille): Try to scroll to the last completed step before.
@@ -110,7 +60,7 @@ const StepsPage = (): React.ReactElement => {
     return (): void => clearTimeout(timeout)
   }, [nextStep])
   const getStepRef = (index: number): undefined|React.RefObject<HTMLDivElement> =>
-    STEPS[index]?.stepId === nextStep ? nextStepRef : undefined
+    Steps[index]?.stepId === nextStep ? nextStepRef : undefined
 
   const onClick = useCallback((page: Page): void => {
     if (!isConnected) {
@@ -134,11 +84,11 @@ const StepsPage = (): React.ReactElement => {
 
   return <Layout>
     <div style={stepsStyle}>
-      {STEPS.map(({isLastStep, title, ...step}, index) => {
-        const stepListJoin = joinList(STEPS.map((step, index) => (index).toString()), t)
+      {Steps.map(({isLastStep, title, ...step}, index) => {
+        const stepListJoin = joinList(Steps.map((step, index) => (index).toString()), t)
         const isDone = !!project.steps?.[step.stepId]?.completed
         const isOpen = !isDone
-          && (!index || !!project.steps?.[STEPS[index - 1].stepId]?.completed)
+          && (!index || !!project.steps?.[Steps[index - 1].stepId]?.completed)
         return <React.Fragment key={index}>
           {index ? <ArrowDownIcon style={arrowStyle} color={colors.SILVER_THREE} /> : null}
           {!isLastStep || isOpen || isDone ?
@@ -147,13 +97,13 @@ const StepsPage = (): React.ReactElement => {
                 {translate(title)}
               </Step>
             </div>
-            : <div style={scrollableStepStyle} ref={getStepRef(STEPS.length)}>
+            : <div style={scrollableStepStyle} ref={getStepRef(Steps.length)}>
               <Button style={interviewStyle} onClick={soonAvailable} type="variable">
                 {t('Préparer un entretien')}
               </Button>
             </div>}
           {!!isLastStep && !isOpen && !isDone ?
-            <Trans count={STEPS.length} style={interviewDisclaimerStyle}>
+            <Trans count={Steps.length} style={interviewDisclaimerStyle}>
               Il est préférable de terminer
               l'étape {{steps: stepListJoin}} avant
               de vous préparer pour les entretiens.
