+++ src/store/partners.ts
@@ -1,5 +1,4 @@
 import Partners from './partners.json'
-import {StepId} from './steps'
 
 export interface Props {
   details?: string
@@ -9,7 +8,7 @@ export interface Props {
   logo: string
   name: string
   partnerId: string
-  steps: readonly StepId[]
+  steps: readonly bayes.maVoie.StepId[]
   title?: string
   url: string
   userCount?: number
@@ -21,7 +20,7 @@ const PartnersWithAssets: readonly Readonly<Props>[] = Partners.
     // Add logos in the /assets/logo folder when adding new ones.
     logo: `${config.canonicalUrl}/assets/logo/${logo}`,
     // TODO(pascal): Try to avoid this type cast.
-    steps: steps as readonly StepId[],
+    steps: steps as readonly bayes.maVoie.StepId[],
   }))
 
 export default PartnersWithAssets
